# internal/apiserver/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.AddPolicyRequest:
    properties:
      action:
        type: string
      changed_by:
        type: string
      reason:
        type: string
      resource_id:
        type: string
      role_id:
        type: string
    required:
    - action
    - resource_id
    - role_id
    type: object
  github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.AssignmentResponse:
    properties:
      granted_by:
        type: string
      id:
        type: string
      role_id:
        type: string
      subject_id:
        type: string
      subject_type:
        type: string
      tenant_id:
        type: string
    type: object
  github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.CheckRequest:
    properties:
      action:
        type: string
      object:
        type: string
      subject_id:
        type: string
      subject_type:
        description: SubjectType 可选：任意主体类型字符串；与 SubjectID 同时省略时使用当前 JWT 用户。
        type: string
    required:
    - action
    - object
    type: object
  github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.CheckResponse:
    properties:
      allowed:
        type: boolean
    type: object
...
  github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.GrantRequest:
    properties:
      granted_by:
        type: string
      role_id:
        type: string
      subject_id:
        type: string
      subject_type:
        enum:
        - user
        type: string
    required:
    - role_id
    - subject_id
    - subject_type
    type: object
...
  github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.PolicyRuleResponse:
    properties:
      action:
        type: string
      domain:
        type: string
      object:
        type: string
      subject:
        type: string
    type: object
...
  github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.RemovePolicyRequest:
    properties:
      action:
        type: string
      changed_by:
        type: string
      reason:
        type: string
      resource_id:
        type: string
      role_id:
        type: string
    required:
    - action
    - resource_id
    - role_id
    type: object
...
  github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.RevokeRequest:
    properties:
      role_id:
        type: string
      subject_id:
        type: string
      subject_type:
        enum:
        - user
        type: string
    required:
    - role_id
    - subject_id
    - subject_type
    type: object
...
host: iam.fangcunmount.cn
info:
  contact:
    email: support@fangcunmount.cn
    name: API Support
    url: https://github.com/FangcunMount/iam-contracts
  description: IAM 系统 REST API 文档，包含认证(Authentication)、授权(Authorization)、身份管理(Identity)和身份提供商(IDP)模块
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://iam.fangcunmount.cn/terms
  title: IAM API Documentation
  version: 1.0.0
paths:
...
  /authz/assignments/{id}:
    delete:
      parameters:
      - description: 分配ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.Response'
      summary: 根据分配ID撤销角色
      tags:
      - Authorization-Assignments
  /authz/assignments/grant:
    post:
      consumes:
      - application/json
      parameters:
      - description: 授予角色请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.GrantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.AssignmentResponse'
              type: object
      summary: 授予角色
      tags:
      - Authorization-Assignments
  /authz/assignments/revoke:
    post:
      consumes:
      - application/json
      parameters:
      - description: 撤销角色请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.RevokeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.Response'
      summary: 撤销角色
      tags:
      - Authorization-Assignments
  /authz/assignments/subject:
    get:
      parameters:
      - description: 主体类型
        enum:
        - user
        - group
        in: query
        name: subject_type
        required: true
        type: string
      - description: 主体ID
        in: query
        name: subject_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.AssignmentResponse'
                  type: array
              type: object
      summary: 列出主体的角色分配
      tags:
      - Authorization-Assignments
  /authz/check:
    post:
      consumes:
      - application/json
      parameters:
      - description: 判定请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.CheckRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.CheckResponse'
              type: object
      summary: 策略判定（Enforce）
      tags:
      - Authorization-Policies
  /authz/policies:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 移除策略请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.RemovePolicyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.Response'
      summary: 移除策略规则
      tags:
      - Authorization-Policies
    post:
      consumes:
      - application/json
      parameters:
      - description: 添加策略请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.AddPolicyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.Response'
      summary: 添加策略规则
      tags:
      - Authorization-Policies
...
  /authz/roles/{id}/assignments:
    get:
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.AssignmentResponse'
                  type: array
              type: object
      summary: 列出角色的分配记录
      tags:
      - Authorization-Assignments
  /authz/roles/{id}/policies:
    get:
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_FangcunMount_iam-contracts_internal_apiserver_interface_authz_restful_dto.PolicyRuleResponse'
                  type: array
              type: object
      summary: 获取角色的策略列表
      tags:
      - Authorization-Policies
...
securityDefinitions:
  BearerAuth:
    description: 'JWT 认证令牌，格式: Bearer {access_token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...